Unify Metadata struct
Merged the FileMetadata and DirectoryMedatata structs
into one general Metadata struct.
Reorganized the filesystem module.
